# project_docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Ozon Price Updater",
        "description": "Система автоматизации мониторинга и корректировки цен на маркетплейсе Ozon",
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/history": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get History",
                "description": "Получение истории изменения цен с возможностью фильтрации.",
                "operationId": "get_history_api_history_get",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Поиск по имени товара",
                            "default": "",
                            "title": "Search"
                        },
                        "description": "Поиск по имени товара"
                    },
                    {
                        "name": "only_changes",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "Только изменения цен",
                            "default": false,
                            "title": "Only Changes"
                        },
                        "description": "Только изменения цен"
                    },
                    {
                        "name": "price_filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Фильтр по типу цены (all/increase/decrease)",
                            "default": "all",
                            "title": "Price Filter"
                        },
                        "description": "Фильтр по типу цены (all/increase/decrease)"
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Период (all/day/week/month)",
                            "default": "all",
                            "title": "Period"
                        },
                        "description": "Период (all/day/week/month)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HistorySchema"
                                    },
                                    "title": "Response Get History Api History Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{product_id}/price": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "Update Product Price",
                "description": "Обновляет цену товара через API Ozon.",
                "operationId": "update_product_price_api_products__product_id__price_post",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Product Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PriceUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/settings": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "Save Settings",
                "description": "Сохраняет настройки приложения.",
                "operationId": "save_settings_api_settings_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "title": "Settings Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/settings": {
            "get": {
                "summary": "Settings Page",
                "description": "Страница настроек.",
                "operationId": "settings_page_settings_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ]
            }
        },
...
        "/history/{product_id}": {
            "get": {
                "summary": "History",
                "description": "Страница истории изменений цен для конкретного товара.",
                "operationId": "history_history__product_id__get",
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Product Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/history": {
            "get": {
                "summary": "Full History",
                "description": "Страница истории изменений цен для всех товаров.",
                "operationId": "full_history_history_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HistorySchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "product_id": {
                        "type": "integer",
                        "title": "Product Id"
                    },
                    "sku": {
                        "type": "string",
                        "title": "Sku"
                    },
                    "product_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Product Name"
                    },
                    "old_price": {
                        "type": "number",
                        "title": "Old Price"
                    },
                    "new_price": {
                        "type": "number",
                        "title": "New Price"
                    },
                    "no_changes": {
                        "type": "boolean",
                        "title": "No Changes",
                        "default": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "product_id",
                    "sku",
                    "old_price",
                    "new_price",
                    "created_at"
                ],
                "title": "HistorySchema",
                "description": "Схема для истории изменения цен"
            },
...
            "PriceUpdateRequest": {
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "title": "Product Id"
                    },
                    "price": {
                        "type": "number",
                        "title": "Price"
                    },
                    "old_price": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Old Price"
                    },
                    "min_price": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Min Price"
                    }
                },
                "type": "object",
                "required": [
                    "product_id",
                    "price"
                ],
                "title": "PriceUpdateRequest",
                "description": "Схема для запроса на обновление цены."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBasic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}